+++ routemaster/logging/__init__.py
@@ -1,0 +1,17 @@
+"""Logging plugin subsystem."""
+
+from routemaster.logging.base import BaseLogger
+from routemaster.logging.plugins import (
+    PluginConfigurationException,
+    register_loggers,
+)
+from routemaster.logging.split_logger import SplitLogger
+from routemaster.logging.python_logger import PythonLogger
+
+__all__ = (
+    'BaseLogger',
+    'SplitLogger',
+    'PythonLogger',
+    'register_loggers',
+    'PluginConfigurationException',
+)
